backend/requirements.txt
gunicorn
django
django-cors-headers
djangorestframework
Pillow
psycopg
python-dotenv
requests
